Player Light Turns off during night
HOTFIX 1 | The player light was being consistently on, so I set a variable in the DayNightCycle.gd script that sets "realtime" to day or night when its day or night. This will turn on abruptly which will be annoying. Probably gonna make an animation player that lasts about 10 minutes (half a day) that slowly turns on the lights? Anyhoo that's it for today :)